last update

+++ libft/ft_bzero.c
@@ -5,16 +5,16 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: ddias-fe <ddias-fe@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/17 16:43:53 by ddias-fe          #+#    #+#             */
-/*   Updated: 2024/01/17 16:43:53 by ddias-fe         ###   ########.fr       */
+/*   Created: 2024/04/09 14:53:49 by ddias-fe          #+#    #+#             */
+/*   Updated: 2024/04/09 14:53:49 by ddias-fe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < n)
